+++ src/handlers/playlist.rs
@@ -62,6 +62,7 @@ pub fn handler(key: Key, app: &mut App) {
       if let (Some(playlists), Some(selected_playlist_index)) =
         (&app.playlists, &app.selected_playlist_index)
       {
+        app.active_playlist_index = Some(selected_playlist_index.to_owned());
         app.track_table.context = Some(TrackTableContext::MyPlaylists);
         app.playlist_offset = 0;
         if let Some(selected_playlist) = playlists.items.get(selected_playlist_index.to_owned()) {