+++ src/ui/audio_analysis.rs
@@ -1,8 +1,10 @@
 use super::util;
 use crate::app::App;
-use rhai::serde::to_dynamic;
-use rhai::{Array, Dynamic, Map};
-use rhai::{Engine, Scope};
+use crate::user_config::VisualStyle;
+use rhai::{
+  serde::{from_dynamic, to_dynamic},
+  Array, Engine, Scope,
+};
 use tui::{
   backend::Backend,
   layout::{Constraint, Direction, Layout},
@@ -11,9 +13,14 @@ use tui::{
   widgets::{BarChart, Block, Borders, Paragraph},
   Frame,
 };
-const PITCHES: [&str; 12] = [
-  "C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B",
-];
+
+// Trait to produce widget
+#[derive(Debug, Clone, serde::Deserialize)]
+struct BarChartInfo {
+  error: bool,
+  labels: Vec<String>,
+  counts: Vec<u64>,
+}
 
 pub fn draw<B>(f: &mut Frame<B>, app: &App)
 where
@@ -35,11 +42,16 @@ where
     .borders(Borders::ALL)
     .border_style(Style::default().fg(app.user_config.theme.inactive));
 
+  let visual_app = app.user_config.get_visualizer_or_default();
   let white = Style::default().fg(app.user_config.theme.text);
   let gray = Style::default().fg(app.user_config.theme.inactive);
-  let width = (chunks[1].width) as f32 / (1 + PITCHES.len()) as f32;
   let tick_rate = app.user_config.behavior.tick_rate_milliseconds;
-  let bar_chart_title = &format!("Pitches | Tick Rate {} {}FPS", tick_rate, 1000 / tick_rate,);
+  let bar_chart_title = &format!(
+    "{} | Tick Rate {} {}FPS",
+    visual_app.name,
+    tick_rate,
+    1000 / tick_rate,
+  );
   let bar_chart_block = Block::default()
     .borders(Borders::ALL)
     .style(white)
@@ -63,7 +75,7 @@ where
     // TODO: Move to own function
     let info: Vec<String> = {
       // TODO: Set up raw engine
-      match &app.visulizer {
+      match &app.visualizer {
         Ok(ast) => {
           let engine = Engine::new();
           match to_dynamic(analysis.clone()) {
@@ -97,58 +109,89 @@ where
       }
     };
 
-    let beat = analysis
-      .beats
-      .iter()
-      .find(|beat| beat.start >= progress_seconds);
+    let texts: Vec<Spans> = info.iter().map(|span| Spans::from(span.clone())).collect();
+    let p = Paragraph::new(texts)
+      .block(analysis_block)
+      .style(Style::default().fg(app.user_config.theme.text));
+    f.render_widget(p, chunks[0]);
 
-    let beat_offset = beat
-      .map(|beat| beat.start - progress_seconds)
-      .unwrap_or(0.0);
+    // TODO: Pass back a struct with render
+    // Struct should have optional error
+    // Match against config
+    // Invalid viz
+    match visual_app.style {
+      VisualStyle::Bar => {
+        let data: Result<Vec<(String, u64)>, String> = {
+          // TODO: Set up raw engine
+          match &app.visualizer {
+            Ok(ast) => {
+              let engine = Engine::new();
+              match to_dynamic(analysis.clone()) {
+                Ok(dynamic_analysis) => {
+                  match engine.call_fn(
+                    &mut Scope::new(),
+                    &ast,
+                    "draw",
+                    (dynamic_analysis, progress_seconds),
+                  ) {
+                    Ok(data) => match from_dynamic(&data) {
+                      Ok(data) => {
+                        let data: BarChartInfo = data;
+                        Ok(
+                          data
+                            .labels
+                            .iter()
+                            .zip(data.counts.iter())
+                            .map(|(label, count)| (label.clone(), count.clone()))
+                            .collect(),
+                        )
+                      }
+                      Err(err) => Err(format!("Unable to type cast: {}", err)),
+                    },
+                    Err(err) => Err(format!("Error in script: {}", err)),
+                  }
+                }
+                Err(err) => Err(format!("Unable to serialize spotify information: {}", err)),
+              }
+            }
+            Err(err) => Err(format!("Compilation error: {}", err)),
+          }
+        };
+        match data {
+          Ok(data) => {
+            let data: Vec<(&str, u64)> =
+              data.iter().map(|item| (item.0.as_str(), item.1)).collect();
+            let width = (chunks[1].width) as f32 / (1 + data.len()) as f32;
 
-    let segment = analysis
-      .segments
-      .iter()
-      .find(|segment| segment.start >= progress_seconds);
-
-    if let Some(segment) = segment {
-      let texts: Vec<Spans> = info.iter().map(|span| Spans::from(span.clone())).collect();
-      let p = Paragraph::new(texts)
-        .block(analysis_block)
-        .style(Style::default().fg(app.user_config.theme.text));
-      f.render_widget(p, chunks[0]);
-
-      let data: Vec<(&str, u64)> = segment
-        .clone()
-        .pitches
-        .iter()
-        .enumerate()
-        .map(|(index, pitch)| {
-          let display_pitch = *PITCHES.get(index).unwrap_or(&PITCHES[0]);
-          let bar_value = ((pitch * 1000.0) as u64)
-            // Add a beat offset to make the bar animate between beats
-            .checked_add((beat_offset * 3000.0) as u64)
-            .unwrap_or(0);
-
-          (display_pitch, bar_value)
-        })
-        .collect();
-
-      let analysis_bar = BarChart::default()
-        .block(bar_chart_block)
-        .data(&data)
-        .bar_width(width as u16)
-        .bar_style(Style::default().fg(app.user_config.theme.analysis_bar))
-        .value_style(
-          Style::default()
-            .fg(app.user_config.theme.analysis_bar_text)
-            .bg(app.user_config.theme.analysis_bar),
-        );
-      f.render_widget(analysis_bar, chunks[1]);
-    } else {
-      f.render_widget(empty_analysis_block(), chunks[0]);
-      f.render_widget(empty_pitches_block(), chunks[1]);
-    };
+            let analysis_bar = BarChart::default()
+              .block(bar_chart_block)
+              .data(data.as_slice())
+              .bar_width(width as u16)
+              .bar_style(Style::default().fg(app.user_config.theme.analysis_bar))
+              .value_style(
+                Style::default()
+                  .fg(app.user_config.theme.analysis_bar_text)
+                  .bg(app.user_config.theme.analysis_bar),
+              );
+            f.render_widget(analysis_bar, chunks[1]);
+          }
+          Err(message) => {
+            let ts: Vec<Spans> = vec![Spans::from(message)];
+            let p = Paragraph::new(ts)
+              .block(bar_chart_block)
+              .style(Style::default().fg(app.user_config.theme.text));
+            f.render_widget(p, chunks[1]);
+          }
+        }
+      }
+      _ => {
+        let ts: Vec<Spans> = vec![Spans::from("Unsupported type.")];
+        let p = Paragraph::new(ts)
+          .block(bar_chart_block)
+          .style(Style::default().fg(app.user_config.theme.text));
+        f.render_widget(p, chunks[1]);
+      }
+    }
   } else {
     f.render_widget(empty_analysis_block(), chunks[0]);
     f.render_widget(empty_pitches_block(), chunks[1]);